Add toast notifications

+++ app/components/LeaveButton.tsx
@@ -2,6 +2,7 @@
 import React from "react";
 import Prisma from "@prisma/client";
 import { useRouter } from "next/navigation";
+import { toastStore } from "../utils/zustand";
 
 type Props = {
   authUser: Prisma.User | null;
@@ -9,15 +10,23 @@ type Props = {
 };
 
 const LeaveButton = (props: Props) => {
-  const { authUser, project } = props;
+  const store = toastStore();
   const router = useRouter();
+
+  const { authUser, project } = props;
+
   const handleLeave = async (e) => {
     e.preventDefault();
+
     const verification = confirm(
       `Are you sure you want to leave project: ${project.name}`,
     );
+
+    if (!verification) return;
+
     try {
       if (!authUser) return;
+
       const response = await fetch(`/api/v1/projects/${project.id}/users`, {
         method: "DELETE",
         headers: {
@@ -25,8 +34,12 @@ const LeaveButton = (props: Props) => {
         },
         body: JSON.stringify({ user: authUser }),
       });
-      if (!response.ok)
+
+      if (!response.ok) {
         throw new Error(`${response.status}: (${response.statusText})`);
+      }
+
+      store.addToast("Project left successfully!");
       router.refresh();
     } catch (e) {
       console.error(e);
